% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.r
\name{build_complex_statespace}
\alias{build_complex_statespace}
\title{Add a statespace to an activity}
\usage{
build_complex_statespace(
  act,
  data,
  statespace0,
  statespace1,
  factors = character(),
  by = character()
)
}
\arguments{
\item{act}{Activity definition}

\item{data}{\code{data.frame} Observed transitions}

\item{statespace0}{\code{data.frame} Possible states for each combination of by variables before transition}

\item{statespace1}{\code{data.frame} Possible states for each combination of by variables after transition}

\item{factors}{\code{character} Variables used in the iterative estimation}

\item{by}{\code{character} Variables not used in the iterative estimation}
}
\value{
Activity definition with statespace
}
\description{
In a complex statespace it is possible to change statespaces with an activity.
Possible states are found using statespace0 and statespace1.
Possible factor combinations are found from data. Thus the data should contain at least one observation for each
interesting combination of factor levels. If pair data is used for estimation it usually is valid also here.
}
