% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example.r
\docType{data}
\name{example}
\alias{example}
\title{Example dataset.}
\format{
A list of data frames:
\describe{
  \item{actprob}{Activity probabilities}
  \item{noman_pairs}{Pair data for growth without management activities}
  \item{thin_pairs}{Pair data for thinning}
  \item{initial_state}{Initial forest state}
  \item{drain_coef}{Coefficient to transform harvested areas into harvest accumulation by timber assortments}
  \item{vol_coef}{Coefficients to transform volume classes into volumes m3/ha}
  \item{income_coef}{Coefficients to transform harvest accumulation into income}
}
}
\usage{
example
}
\description{
A list containing the necessary datasets for EFDM forest scenario example.
}
\keyword{datasets}
