% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.r
\name{extract_transitions}
\alias{extract_transitions}
\title{Extract transitions from an activity}
\usage{
extract_transitions(act)
}
\arguments{
\item{act}{Activity}
}
\value{
\code{data.frame} where prob is the transition probability from current
state (with suffix 0) to next state (with suffix 1).
}
\description{
Extract transitions from an activity
}
