% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example.r
\docType{data}
\name{MetsaKasvVyoh}
\alias{MetsaKasvVyoh}
\title{Finnish bio-geographical regions}
\format{
An sf-object:
\describe{
  \item{region}{Factor with three levels: North, Middle, South}
  \item{geometry}{Polygon, each region is composed of multiple polygons}
}
}
\usage{
MetsaKasvVyoh
}
\description{
A low resolution copy of the finnish bio-geographical regions.
The original shapefile was provided by Finnish Environment Institute.
See \url{https://ckan.ymparisto.fi/fi/dataset/metsakasvillisuusvyohykkeet}.
}
\keyword{datasets}
