% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.r
\name{prior_grow}
\alias{prior_grow}
\title{Prior for deterministic growth}
\usage{
prior_grow(variable, howmuch = 1)
}
\arguments{
\item{variable}{Name of the variable to grow}

\item{howmuch}{Amount of growth}
}
\description{
Prior for deterministic growth
}
