% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.r
\name{build_complex_statespace}
\alias{build_complex_statespace}
\title{Add a statespace to an activity}
\usage{
build_complex_statespace(
  act,
  statespace0,
  statespace1,
  factors = character(),
  by = character()
)
}
\arguments{
\item{act}{Activity definition}

\item{statespace0}{\code{data.frame} Statespace before transition}

\item{statespace1}{\code{data.frame} Statespace after transition}

\item{factors}{\code{character} Variables used by the activity}

\item{by}{\code{character} Variables that split the statespace}
}
\value{
Activity definition with statespace
}
\description{
Add a statespace to an activity
}
\details{
In a complex statespace it is possible to change statespaces with an activity.
Since statespace is the collection of classes of variables this means that the
classification changes.
}
\seealso{
build_state_space
}
