% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.r
\name{build_statespace}
\alias{build_statespace}
\title{Add a statespace to an activity}
\usage{
build_statespace(act, statespace, factors = character(), by = character())
}
\arguments{
\item{act}{Activity definition}

\item{statespace}{\code{data.frame}}

\item{factors}{\code{character} Variables used in the iterative estimation}

\item{by}{\code{character} Variables not used in the iterative estimation}
}
\value{
Activity definition with statespace
}
\description{
Statespace is the collection of levels for each state variable.
}
\details{
If statespace is changing as a result of the activity see \code{\link{build_complex_statespace}}.
}
\examples{

statespacepine <- expand.grid(species="pine", vol=1:5, age=1:3, stringsAsFactors=FALSE)
statespacespruce <- expand.grid(species="spruce", vol=1:4, age=1:3, stringsAsFactors=FALSE)
statespace <- rbind(statespacepine, statespacespruce)
act <- define_activity("nomanagement", c("vol", "age"))
act <- build_statespace(act, statespace, by="species")

\dontshow{
statespace <- expand.grid(ds=c("sp", "pi"), region=c("n", "s"), vol=1:3, stringsAsFactors=FALSE)
act <- define_activity("test", c("vol"))
act <- build_statespace(act, subset(statespace, !(ds=="sp"&region=="n")), factors=c("ds", "region"))
stopifnot(nrow(act$statespace[[1]])==3)

statespace <- expand.grid(a=1:2, b=c("f", "g"), c=c(4,9), vol=1:3, stringsAsFactors=FALSE)
ii <- sample(nrow(statespace), sample(nrow(statespace), 1))
act <- define_activity("test", c("vol"))
act <- build_statespace(act, statespace[ii,], factors=c("a", "b"), by="c")
if(sum(sapply(act$statespace, nrow))!=nrow(unique(statespace[ii,c("a", "b", "c")]))) {
  stop(paste("build_statespace failed", list(ii)))
}
}
}
