% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.r
\name{define_activity}
\alias{define_activity}
\title{Define activity}
\usage{
define_activity(name, statespace, probname = name)
}
\arguments{
\item{name}{Name of activity used in reporting}

\item{statespace}{Names of variables where changes happen}

\item{probname}{Name of activity in activity probabilities data}
}
\description{
Define activity
}
\examples{
define_activity("nomanagement", c("vol", "age"))
}
