% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim.r
\name{estimatetransprobs}
\alias{estimatetransprobs}
\title{Estimate Transition Probabilities from Pairdata}
\usage{
estimatetransprobs(act, pairdata, prior)
}
\arguments{
\item{act}{Activity definition with statespace}

\item{pairdata}{\code{data.frame} Observed transitions}

\item{prior}{function or character}
}
\value{
Activity definition with transition probabilities
}
\description{
The estimation uses an iterative Bayesian algorithm that is explained in
\url{https://github.com/ec-jrc/efdm/blob/master/documents/EFDMinstructions/Seija_Mathematics_behind_EFDM.pdf}.
}
\details{
If pairdata is NULL prior is used by itself.
}
