% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.r
\name{prior_ff}
\alias{prior_ff}
\title{Prior for final felling}
\usage{
prior_ff()
}
\description{
Prior for final felling
}
\examples{
statespace <- expand.grid(a=1:2, b=1:2, vol=1:15, age=1:35)
act <- define_activity("test", c("vol", "age"))
act <- build_statespace(act, statespace, by=c("a", "b"))
act1 <- estimatetransprobs(act, NULL, prior_ff())
}
