% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{runEFDM}
\alias{runEFDM}
\title{Run Markov Forest Dynamics Model}
\usage{
runEFDM(state0, actprob, activities, n)
}
\arguments{
\item{state0}{\code{data.frame} Initial state}

\item{actprob}{\code{data.frame} Activation probabilities}

\item{activities}{\code{list} A list of activities}

\item{n}{\code{integer} Number of time steps required}
}
\value{
\code{data.frame} State of each time step
}
\description{
Run Markov Forest Dynamics Model
}
